create definitions.ts and extract values from artifact-downloader
