+++ src/definitions.ts
@@ -1,0 +1,11 @@
+const RAILGUN_ARTIFACTS_CID_ROOT =
+  'QmeBrG7pii1qTqsn7rusvDiqXopHPjCT9gR4PsmW7wXqZq'
+
+enum ArtifactName {
+  ZKEY = 'zkey',
+  WASM = 'wasm',
+  VKEY = 'vkey',
+  DAT = 'dat',
+}
+
+export { RAILGUN_ARTIFACTS_CID_ROOT, ArtifactName }
